Adapted to Pos x and y being i16 (allow off screen position).

+++ src/ui/screen.rs
@@ -14,8 +14,8 @@ pub fn build<'a>() -> Element<AppState> {
         Element::new(Pos::new(0, 0), Look::from(make_grid(columns(), rows())));
 
     screen.listener.on_loop = Some(Box::new(|_el, state, _event| {
-        let x = columns().saturating_sub(APP_WIDTH) / 2;
-        let y = rows().saturating_sub(APP_HEIGHT) / 2;
+        let x = (columns().saturating_sub(APP_WIDTH) / 2) as i16;
+        let y = (rows().saturating_sub(APP_HEIGHT) / 2) as i16;
 
         if x != state.app_x || y != state.app_y {
             go_to(0, 1);
