+++ src/ui/splash_footer.rs
@@ -5,13 +5,13 @@ pub fn build<'a>() -> Element<SplashState> {
     let mut splash_footer = Element::new(Pos::new(0, 0), Look::new());
 
     splash_footer.listener.on_state = Some(Box::new(|el, _event| {
-        let term_cols = columns();
-        let term_rows = rows();
+        let term_cols = columns() as i16;
+        let term_rows = rows() as i16;
 
         let text = "Made with Rust";
         el.pos
             .x
-            .set((term_cols.saturating_sub(text.len() as u16)) / 2);
+            .set((term_cols.saturating_sub(text.len() as i16)) / 2);
         el.pos.y.set(term_rows.saturating_sub(1));
         el.look
             .update(terminal_style::format::bold(terminal_style::format::faint(
